# 89 rewrite user profile # 88 office manager # 86 fixed bad email # 84 password is null refactoring

+++ src/main/java/defencer/controller/entity/InstructorController.java
@@ -148,7 +148,10 @@ public class InstructorController implements Initializable {
             NotificationUtil.warningAlert("Warning", "Select instructor firstly", NotificationUtil.SHORT);
             return;
         }
-
+        if (!"FREE".equals(instructor.getStatus())) {
+            NotificationUtil.warningAlert("Warning", "Unfortunately you can't this user because he has a project", NotificationUtil.SHORT);
+            return;
+        }
         try {
             ServiceFactory.getInstructorService().deleteEntity(instructor);
             observableInstructors.clear();
